alteracao funcoes arquivo 2
